chore: convert dateRange format to pipe
